+++ Modern JavaScript From The Beginning/Section 8: API Projects/78 - 81. WeatherJS/storage.js
@@ -1,0 +1,23 @@
+const defaultCountry = 'Norway';
+
+class Storage {
+    constructor () {
+        this.country;
+    }
+
+    getCountryData () {
+        if ('country' in localStorage) {
+            this.country = localStorage.getItem('country');
+        } else {
+            this.country = defaultCountry;
+        }
+
+        return {
+            country: this.country
+        };
+    }
+
+    setCountryData (country) {
+        localStorage.setItem('country', country);
+    }
+}